+++ docs/md/releasenotes.md
@@ -7,6 +7,7 @@
 - Improve performance of `ClassPool.removeClass`.
 - Allow more configuration of `ExecutingInvocationUnit` by using `ExecutingInvocationUnit.Builder`.
 - Add a mode to `ExecutingInvocationUnit` to approximate reference of types not supported for the execution via reflection.
+- Implement `Autocloseable` in `DataEntryWriter` interface.
 
 ### Bug fixes